+++ src/huffman/huffman/io/binary-io.cpp
@@ -1,8 +1,7 @@
 #include "io/binary-io.h"
-#include <iostream>
 
 
-u64 io::read_bits(unsigned nbits, io::InputStream& input) {
+u64 io::read_bits(u64 nbits, io::InputStream& input) {
 	u64 result = 0;
 
 	while (nbits-- != 0) {
@@ -14,7 +13,7 @@ u64 io::read_bits(unsigned nbits, io::InputStream& input) {
 }
 
 // Waarom falen de tests die meer dan 8 bits nodig hebben als de bitwise AND "& 1" er niet staat?
-void io::write_bits(u64 value, unsigned nbits, io::OutputStream& output) {
+void io::write_bits(u64 value, u64 nbits, io::OutputStream& output) {
 	while (nbits-- != 0) {
 		output.write((value >> nbits) & 1);
 	}
