+++ src/huffman/huffman/io/binary-io.h
@@ -6,8 +6,8 @@
 
 
 namespace io {
-	u64 read_bits(unsigned, io::InputStream&);
-	void write_bits(u64, unsigned, io::OutputStream&);
+	u64 read_bits(u64, io::InputStream&);
+	void write_bits(u64, u64, io::OutputStream&);
 }
 
 #endif
